// Application.Client/src/app/views/sales/models/sale-invoice/sale-invoice-response-dto.model.ts
import { Customer } from "app/views/configuration/models/customer/customer.model";
import { ProductView } from "app/views/configuration/models/product/product-view.model";
import { Store } from "app/views/configuration/models/store/store.model";

export interface SaleInvoiceResponseDTO {
  id?: string;
  invoiceNo: string;
  invoiceDate: string;
  deliveryNoteNo: string;
  saleOrderNo: string;
  orderDate: string;
  customerId: string;
  customerMarketingOfficerId?: string;
  customerTerritoryId: string;
  subtotal: number;
  totalPercentageDiscountAmount: number;
  discount: number;
  offerDiscount: number;
  otherDiscount: number;
  total: number;
  transportationCost: number;
  depoCharge: number;
  netTotal: number;
  referenceNo: string;
  bankDetails: string;
  termAndCondition: string;
  status: number;
  statusName: string;
  remark: string;
  paymentTerm: number;
  paymentTermName: string;
  moneyReceiptNo?: string;
  storeId?: string;
  transport: number;
  checkedBy: string;
  approvedBy: string;
  createdOn: string;
  updatedOn: string;
  createdBy: string;
  updatedBy: string;
  financialYearId: string;
  customer: Customer;
  store: Store;
  saleInvoiceDetails: SaleInvoiceResponseDetail[];
}
export interface SaleInvoiceResponseDetail {
  id?: string;
  saleInvoiceId?: string;
  saleOrderDetailId?: string;
  productId: string;
  measurementUnitName?: string;
  primaryQuantity: number;
  quantity: number;
  primaryBonusQuantity: number;
  bonusQuantity: number;
  rate: number;
  netRate: number;
  discountPerUnit: number;
  offerDiscountPerUnit: number;
  discountAmount: number;
  invoiceDiscountPerUnit: number;
  cashDiscountPerUnit: number;
  specialDiscountPerUnit: number;
  vatPercentage: number;
  discountPercentage: number;
  percentageDiscountAmount: number;
  amount: number;
  deliveryNoteNo: string;
  deliveryDate: string;
  deliveryPlace: string;
  otherDiscountPerUnit: number;
  transportationCostPerUnit: number;
  depoChargePerUnit: number;
  product?: ProductView;
}